requirements_gpu.txt
librosa>=0.5
tensorflow-gpu>=1.0.0
pytube>=6.2.2
moviepy>=0.2.2.13
Pillow>=3.1.0